+++ src/test/java/de/gematik/demis/nrs/rules/RulesServiceOtherIT.java
@@ -38,32 +38,17 @@ import java.nio.file.Path;
 import java.util.Optional;
 import java.util.stream.Stream;
 import org.hl7.fhir.r4.model.Bundle;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-/**
- * This class uses the routingConfig_73enabled.json configuration for it's test cases. It can be
- * renamed once 7.3 is ready.
- */
-@SpringBootTest(
-    classes = NotificationRoutingApplication.class,
-    useMainMethod = SpringBootTest.UseMainMethod.ALWAYS,
-    webEnvironment = SpringBootTest.WebEnvironment.MOCK,
-    properties = {
-      "nrs.routing-rules=rules/routingConfig_73enabled.json",
-      "nrs.lookup-data-directory=src/test/resources/integrationtest/data/lookup",
-      "nrs.rules-start-id=start"
-    })
-class RulesService73IT {
+/** Verify other routing configurations behind feature flags */
+class RulesServiceOtherIT {
 
   private static final String BASE_PATH = "src/test/resources/fhir";
-
-  @Autowired private RulesService rulesService;
-
-  @Autowired private FhirContext fhirContext;
 
   private static Stream<Arguments> bundleToExpectedResultId() {
     return Stream.of(
@@ -82,24 +67,66 @@ class RulesService73IT {
         arguments(
             Path.of(BASE_PATH, "7_3/disease-anonymous-nonnominal.json"), "disease_7_3_anonymous"),
         arguments(Path.of(BASE_PATH, "7_4/negative-covid19-bundle.json"), "laboratory_7_4"),
-        arguments(Path.of(BASE_PATH, "invalid_71.json"), "laboratory_7_1_anonymous_follow_up"),
+        arguments(Path.of(BASE_PATH, "invalid_71.json"), "laboratory_7_1"),
         arguments(
-            Path.of(BASE_PATH, "invalid_71_multiple_observations_mixed.json"),
-            "laboratory_7_1_anonymous_follow_up"),
+            Path.of(BASE_PATH, "invalid_71_multiple_observations_mixed.json"), "laboratory_7_1"),
         arguments(
-            Path.of(BASE_PATH, "invalid_71_multiple_observations_all_neg.json"),
-            "laboratory_7_1_anonymous_follow_up"));
+            Path.of(BASE_PATH, "invalid_71_multiple_observations_all_neg.json"), "laboratory_7_1"));
   }
 
-  @ParameterizedTest
-  @MethodSource("bundleToExpectedResultId")
-  void thatRuleMatches(final Path path, final String expectedRule) throws IOException {
-    final String bundleJson = Files.readString(path);
-    final Bundle bundle = (Bundle) fhirContext.newJsonParser().parseResource(bundleJson);
+  @SpringBootTest(
+      classes = NotificationRoutingApplication.class,
+      useMainMethod = SpringBootTest.UseMainMethod.ALWAYS,
+      webEnvironment = SpringBootTest.WebEnvironment.MOCK,
+      properties = {
+        "nrs.routing-rules=rules/routingConfig_73enabled.json",
+        "nrs.lookup-data-directory=src/test/resources/integrationtest/data/lookup",
+        "nrs.rules-start-id=start"
+      })
+  @Nested
+  class For73 {
+    @Autowired private RulesService rulesService;
 
-    final Optional<Result> resultCandidate = rulesService.evaluateRules(bundle);
-    assertThat(resultCandidate).isPresent();
-    final Result result = resultCandidate.get();
-    assertThat(result.id()).isEqualTo(expectedRule);
+    @Autowired private FhirContext fhirContext;
+
+    @MethodSource("de.gematik.demis.nrs.rules.RulesServiceOtherIT#bundleToExpectedResultId")
+    @ParameterizedTest
+    void thatRuleMatches(final Path path, final String expectedRule) throws IOException {
+      final String bundleJson = Files.readString(path);
+      final Bundle bundle = (Bundle) fhirContext.newJsonParser().parseResource(bundleJson);
+
+      final Optional<Result> resultCandidate = rulesService.evaluateRules(bundle);
+      assertThat(resultCandidate).isPresent();
+      final Result result = resultCandidate.get();
+      assertThat(result.id()).isEqualTo(expectedRule);
+    }
+  }
+
+  @SpringBootTest(
+      classes = NotificationRoutingApplication.class,
+      useMainMethod = SpringBootTest.UseMainMethod.ALWAYS,
+      webEnvironment = SpringBootTest.WebEnvironment.MOCK,
+      properties = {
+        "nrs.routing-rules=rules/routingConfig_with_follow_up.json",
+        "nrs.lookup-data-directory=src/test/resources/integrationtest/data/lookup",
+        "nrs.rules-start-id=start"
+      })
+  @Nested
+  class FollowUp {
+    @Autowired private RulesService rulesService;
+
+    @Autowired private FhirContext fhirContext;
+
+    @ParameterizedTest
+    @MethodSource("de.gematik.demis.nrs.rules.RulesServiceOtherIT#bundleToExpectedResultId")
+    void thatRuleMatches(final Path path, final String expectedRule) throws IOException {
+      final String bundleJson = Files.readString(path);
+      final Bundle bundle = (Bundle) fhirContext.newJsonParser().parseResource(bundleJson);
+
+      final Optional<Result> resultCandidate = rulesService.evaluateRules(bundle);
+      assertThat(resultCandidate).isPresent();
+      final Result result = resultCandidate.get();
+      assertThat(result.id()).isEqualTo(expectedRule);
+    }
   }
 }